chore: run tests for `noir_wasm` in node
