+++ compiler/wasm/package.json
@@ -20,8 +20,8 @@
   },
   "scripts": {
     "build": "bash ./build.sh",
-    "test": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha",
-    "test:node": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\" }' mocha",
+    "test": "yarn test:node && yarn test:browser",
+    "test:node": "env TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' mocha",
     "test:browser": "web-test-runner",
     "clean": "chmod u+w web nodejs || true && rm -rf ./nodejs ./web ./target ./result",
     "publish": "echo 📡 publishing `$npm_package_name` && yarn npm publish",
