+++ compiler/wasm/test/browser/index.test.ts
@@ -1,6 +1,7 @@
 import { expect } from '@esm-bundle/chai';
-import initNoirWasm from '@noir-lang/noir_wasm';
-import { compileNoirSource, nargoArtifactPath, noirSourcePath } from '../shared';
+import initNoirWasm, { compile } from '@noir-lang/noir_wasm';
+import { initializeResolver } from '@noir-lang/source-resolver';
+import { nargoArtifactPath, noirSourcePath } from '../shared';
 
 beforeEach(async () => {
   await initNoirWasm();
@@ -20,6 +21,35 @@ async function getSource(): Promise<string> {
 async function getPrecompiledSource(): Promise<any> {
   const compiledData = await getFileContent(nargoArtifactPath);
   return JSON.parse(compiledData);
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export async function compileNoirSource(noir_source: string): Promise<any> {
+  console.log('Compiling Noir source...');
+
+  initializeResolver((id: string) => {
+    console.log(`Resolving source ${id}`);
+
+    const source = noir_source;
+
+    if (typeof source === 'undefined') {
+      throw Error(`Could not resolve source for '${id}'`);
+    } else if (id !== '/main.nr') {
+      throw Error(`Unexpected id: '${id}'`);
+    } else {
+      return source;
+    }
+  });
+
+  try {
+    const compiled_noir = compile('main.nr');
+
+    console.log('Noir source compilation done.');
+
+    return compiled_noir;
+  } catch (e) {
+    console.log('Error while compiling:', e);
+  }
 }
 
 describe('noir wasm compilation', () => {
