--- compiler/wasm/test/index.d.ts
@@ -1,1 +1,0 @@
-declare module '@noir-lang/source-resolver';
